Support for controlling the FCCDicts dir with env FCCDICTSDIR
FCCDICTSDIR should point to a directory containing the json files
and the rest.
For example, on lxplus:

$ export FCCDICTSDIR=/afs/cern.ch/work/h/helsens/public/FCCDicts/

(the trailing '/' is optional, it will be checked and added inside)

The default is called 'deffccdicts' read from the new file common_default.py .
It is currently set to '/afs/cern.ch/work/h/helsens/public/FCCDicts/' , so nothing
should change for people able to address that directory.